export LabwareDetailOverflowMenu and migrate css modules
