+++ app/src/organisms/Desktop/ProtocolDetails/ProtocolLabwareDetails/index.tsx
@@ -1,5 +1,3 @@
-import { useState } from 'react'
-import { createPortal } from 'react-dom'
 import { useTranslation } from 'react-i18next'
 
 import {
@@ -10,23 +8,16 @@ import {
   Flex,
   Icon,
   InfoScreen,
-  MenuItem,
-  NO_WRAP,
-  OverflowBtn,
-  POSITION_ABSOLUTE,
-  POSITION_RELATIVE,
   SPACING,
   StyledText,
-  useMenuHandleClickOutside,
 } from '@opentrons/components'
 import { getLabwareDefIsStandard } from '@opentrons/shared-data'
 
-import { getTopPortalEl } from '/app/App/portal'
 import { Divider } from '/app/atoms/structure'
-import { LabwareDetails } from '/app/organisms/Desktop/Labware/LabwareDetails'
 import { getRequiredLabwareDetailsFromLoadCommands } from '/app/transformations/commands'
 
-import type { MouseEventHandler } from 'react'
+import { LabwareDetailOverflowMenu } from './LabwareDetailOverflowMenu'
+
 import type { RunTimeCommand } from '@opentrons/shared-data'
 import type { LabwareDefAndDate } from '/app/local-resources/labware'
 
@@ -141,73 +132,5 @@ export const ProtocolLabwareDetailItem = (
         <LabwareDetailOverflowMenu labware={labware} />
       </Flex>
     </>
-  )
-}
-
-interface LabwareDetailOverflowMenuProps {
-  labware: LabwareDefAndDate
-}
-
-export const LabwareDetailOverflowMenu = (
-  props: LabwareDetailOverflowMenuProps
-): JSX.Element => {
-  const { labware } = props
-  const { t } = useTranslation('protocol_details')
-  const {
-    menuOverlay,
-    handleOverflowClick,
-    showOverflowMenu,
-    setShowOverflowMenu,
-  } = useMenuHandleClickOutside()
-  const [showLabwareDetailSlideout, setShowLabwareDetailSlideout] =
-    useState<boolean>(false)
-
-  const handleClickMenuItem: MouseEventHandler<HTMLButtonElement> = e => {
-    e.preventDefault()
-    setShowOverflowMenu(false)
-    setShowLabwareDetailSlideout(true)
-  }
-  return (
-    <Flex
-      flexDirection={DIRECTION_COLUMN}
-      position={POSITION_RELATIVE}
-      marginRight={SPACING.spacing8}
-      marginLeft={SPACING.spacingAuto}
-    >
-      <Flex>
-        <OverflowBtn onClick={handleOverflowClick} />
-      </Flex>
-      {showOverflowMenu ? (
-        <Flex
-          whiteSpace={NO_WRAP}
-          zIndex={10}
-          borderRadius="4px 4px 0px 0px"
-          boxShadow="0px 1px 3px rgba(0, 0, 0, 0.2)"
-          position={POSITION_ABSOLUTE}
-          backgroundColor={COLORS.white}
-          top="2.3rem"
-          right={0}
-          flexDirection={DIRECTION_COLUMN}
-        >
-          <MenuItem onClick={handleClickMenuItem}>
-            {t('go_to_labware_definition')}
-          </MenuItem>
-        </Flex>
-      ) : null}
-      {createPortal(
-        <>
-          {menuOverlay}
-          {showLabwareDetailSlideout ? (
-            <LabwareDetails
-              labware={labware}
-              onClose={() => {
-                setShowLabwareDetailSlideout(false)
-              }}
-            />
-          ) : null}
-        </>,
-        getTopPortalEl()
-      )}
-    </Flex>
   )
 }
