+++ app/src/organisms/Desktop/ProtocolDetails/ProtocolLabwareDetails/labwareretailoverflowmenu.module.css
@@ -1,0 +1,20 @@
+.menu_item_container {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  margin-right: var(--spacing-8);
+  margin-left: auto;
+}
+
+.menu_item_wrapper {
+  position: absolute;
+  z-index: 10;
+  top: 2.3rem;
+  right: 0;
+  display: flex;
+  flex-direction: column;
+  border-radius: 4px 4px 0 0;
+  background-color: var(--white);
+  box-shadow: 0 1px 3px rgb(0 0 0 / 20%);
+  white-space: nowrap;
+}
